auth_api/python/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
PyJWT==2.1.0
